fix: logout button in navbar
